+++ common/src/main/java/io/github/flemmli97/flan/mixin/PlayerClaimMixin.java
@@ -3,6 +3,7 @@ package io.github.flemmli97.flan.mixin;
 import com.mojang.authlib.GameProfile;
 import io.github.flemmli97.flan.claim.Claim;
 import io.github.flemmli97.flan.event.EntityInteractEvents;
+import io.github.flemmli97.flan.platform.ClaimEvents;
 import io.github.flemmli97.flan.player.PlayerClaimData;
 import io.github.flemmli97.flan.utils.IPlayerClaimImpl;
 import io.github.flemmli97.flan.utils.VanillaFlightStateTracker;
@@ -12,6 +13,7 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -19,6 +21,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(ServerPlayer.class)
 public abstract class PlayerClaimMixin extends Player implements IPlayerClaimImpl, VanillaFlightStateTracker {
+
+    @Shadow
+    protected abstract void nextContainerCounter();
 
     @Unique
     private PlayerClaimData flan$ClaimData;
@@ -43,7 +48,11 @@ public abstract class PlayerClaimMixin extends Player implements IPlayerClaimImp
         if (this.flan$otherFlightState != this.getAbilities().mayfly && this.flan$enabledFlight) {
             this.flan$otherFlightState = this.getAbilities().mayfly;
         }
-        this.flan$CurrentClaim = EntityInteractEvents.currentClaimTick((ServerPlayer) (Object) this, this.flan$CurrentClaim);
+        Claim newClaim = EntityInteractEvents.currentClaimTick((ServerPlayer) (Object) this, this.flan$CurrentClaim);
+        if (this.flan$CurrentClaim != newClaim)
+            ClaimEvents.INSTANCE.borderCross((ServerPlayer) (Object) this, newClaim, this.flan$CurrentClaim);
+        this.flan$CurrentClaim = newClaim;
+
         this.flan$ClaimData.tick(this.flan$CurrentClaim);
     }