+++ neoforge/src/main/java/io/github/flemmli97/flan/forge/platform/ClaimEventsImpl.java
@@ -1,0 +1,27 @@
+package io.github.flemmli97.flan.forge.platform;
+
+import io.github.flemmli97.flan.api.forge.ClaimBorderCrossEvent;
+import io.github.flemmli97.flan.api.forge.PermissionCheckEvent;
+import io.github.flemmli97.flan.claim.Claim;
+import io.github.flemmli97.flan.platform.ClaimEvents;
+import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionResult;
+import net.neoforged.neoforge.common.NeoForge;
+import org.jetbrains.annotations.Nullable;
+
+public class ClaimEventsImpl implements ClaimEvents {
+
+    @Override
+    public InteractionResult claimCheck(ServerPlayer player, ResourceLocation permission, BlockPos pos) {
+        PermissionCheckEvent event = new PermissionCheckEvent(player, permission, pos);
+        NeoForge.EVENT_BUS.post(event);
+        return event.getActionResult();
+    }
+
+    @Override
+    public void borderCross(ServerPlayer player, @Nullable Claim enter, @Nullable Claim exit) {
+        NeoForge.EVENT_BUS.post(new ClaimBorderCrossEvent(player, enter, exit));
+    }
+}